codes: move to codes.go and romeve not JSON-RPC 2 specific codes
