+++ errors.go
@@ -64,36 +64,7 @@ var _ error = (*constErr)(nil)
 // Error implements error.Error.
 func (e constErr) Error() string { return string(e) }
 
-// This file contains the Go forms of the wire specification.
-//
-// See http://www.jsonrpc.org/specification for details.
-//
-// list of JSON-RPC errors.
-var (
-	// ErrUnknown should be used for all non coded errors.
-	ErrUnknown = NewError(UnknownError, "JSON-RPC unknown error")
-
-	// ErrParse is used when invalid JSON was received by the server.
-	ErrParse = NewError(ParseError, "JSON-RPC parse error")
-
-	// ErrInvalidRequest is used when the JSON sent is not a valid Request object.
-	ErrInvalidRequest = NewError(InvalidRequest, "JSON-RPC invalid request")
-
-	// ErrMethodNotFound should be returned by the handler when the method does
-	// not exist / is not available.
-	ErrMethodNotFound = NewError(MethodNotFound, "JSON-RPC method not found")
-
-	// ErrInvalidParams should be returned by the handler when method
-	// parameter(s) were invalid.
-	ErrInvalidParams = NewError(InvalidParams, "JSON-RPC invalid params")
-
-	// ErrInternal is not currently returned but defined for completeness.
-	ErrInternal = NewError(InternalError, "JSON-RPC internal error")
-
-	// ErrServerOverloaded is returned when a message was refused due to a
-	// server being temporarily unable to accept any new messages.
-	ErrServerOverloaded = NewError(ServerOverloaded, "JSON-RPC overloaded")
-
+const (
 	// ErrIdleTimeout is returned when serving timed out waiting for new connections.
 	ErrIdleTimeout = constErr("timed out waiting for new connections")
 )
